+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add upload_options plugin for dynamically generating storage-specific upload options (janko-m)
+
 * Allow the context hash to be modified (janko-m)
 
 * Fix extension not being returned for storages which remove it from ID (Flickr, SQL, GridFS) (janko-m)